move SHARE prices from models to settings
