+++ tapir/settings.py
@@ -11,6 +11,7 @@ https://docs.djangoproject.com/en/3.1/ref/settings/
 """
 import os
 import sys
+from decimal import Decimal
 from pathlib import Path
 
 import celery.schedules
@@ -243,3 +244,6 @@ if ENABLE_SILK_PROFILING:
     SILKY_PYTHON_PROFILER = True
     SILKY_PYTHON_PROFILER_BINARY = True
     SILKY_META = True
+
+COOP_SHARE_PRICE = Decimal(100)
+COOP_ENTRY_AMOUNT = Decimal(10)
